Renamed a config and discovery packages to internal in the SamsungTV binding. See openhab/static-code-analysis#100
