+++ addons/binding/org.openhab.binding.samsungtv/src/main/java/org/openhab/binding/samsungtv/internal/config/SamsungTvConfiguration.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.samsungtv.config;
+package org.openhab.binding.samsungtv.internal.config;
 
 /**
  * Configuration class for {@link SamsungTvBinding} device.
